+++ packages/documentation-v6/cypress/snapshots/styles/headings.snapshot.ts
@@ -1,6 +1,0 @@
-describe('Headings', () => {
-  it('default', () => {
-    cy.visit('/iframe.html?id=hidden-headings-snapshot-test--page');
-    cy.percySnapshot('Headings', { widths: [1024] });
-  });
-});