+++ packages/documentation/src/stories/internet-header/components/footer/custom-config/footer-config.stories.ts
@@ -1,0 +1,5 @@
+import * as FooterStories from '../footer.stories';
+
+export default { ...FooterStories.default, title: 'Internet Header/Footer/Custom Config' };
+
+export const Default = FooterStories.CustomConfig;