+++ packages/documentation/src/stories/internet-header/components/header/overrides-stories/header-os-flyout-overrides.stories.ts
@@ -1,0 +1,8 @@
+import * as HeaderStories from '../header.stories';
+
+export default {
+  ...HeaderStories.default,
+  title: 'Internet Header/Header/Os Flyout Overrides',
+};
+
+export const Default = HeaderStories.OSFlyoutOverrides;