+++ packages/documentation/src/stories/internet-header/internet-header-lazy-loaded.sample.js
@@ -1,0 +1,4 @@
+// main.ts / index.js / ...
+import { defineCustomElements } from '@swisspost/internet-header/loader';
+
+defineCustomElements();