+++ packages/documentation/src/utils/inputComponentsGetCombinations.ts
@@ -1,35 +1,39 @@
-const SHORTHINT =
+
+const SHORT_HINT =
   'Hintus ipsum dolor sit amet consectetur adipisicing elit. Vero mollitia magnam quo quam saepe. Aliquam tempore non deleniti culpa reprehenderit.';
-const LONGHINT =
+const LONG_HINT =
   'Hintus ipsum dolor sit amet consectetur adipisicing elit. Voluptatem maxime eius aut quae ducimus dignissimos pariatur suscipit distinctio, accusamus laudantium, sint quibusdam nisi optio? Ut quae obcaecati, harum ullam quos beatae, ipsam enim, placeat eligendi dolores excepturi. Quia quod eligendi ab voluptas modi id distinctio iure vel possimus deserunt, amet, dolores laboriosam quas qui aut laborum? Et numquam esse laboriosam totam quod sapiente recusandae consectetur optio, quaerat quia.';
-const SHORTLABEL = 'Label';
-const LONGTEXT =
+const SHORT_LABEL = 'Label';
+const LONG_TEXT =
+
   'Lorem ipsum dolor sit amet consectetur adipisicing elit. Voluptatem maxime eius aut quae ducimus dignissimos pariatur suscipit distinctio, accusamus laudantium, sint quibusdam nisi optio? Ut quae obcaecati, harum ullam quos beatae, ipsam enim, placeat eligendi dolores excepturi';
 
 export const COMBINATIONS = [
   {
     title: true, // This property is true when a heading should be rendered above the story
-    label: `${SHORTLABEL} - no Hint`,
+
+    label: `${SHORT_LABEL} - no Hint`,
     hint: null,
   },
   {
-    label: `${SHORTLABEL} - short Hint`,
-    hint: SHORTHINT,
+    label: `${SHORT_LABEL} - short Hint`,
+    hint: SHORT_HINT,
   },
   {
-    label: `$Label - long - ${LONGTEXT}`,
-    hint: LONGHINT,
+    label: `${SHORT_LABEL} - long - ${LONG_TEXT}`,
+    hint: LONG_HINT,
   },
   {
-    label: `${SHORTLABEL} - Disabled`,
+    label: `${SHORT_LABEL} - Disabled`,
     disabled: true,
   },
   {
-    label: `${SHORTLABEL} - Valid`,
+    label: `${SHORT_LABEL} - Valid`,
     validation: 'is-valid',
   },
   {
-    label: `${SHORTLABEL} - Invalid`,
+    label: `${SHORT_LABEL} - Invalid`,
+
     validation: 'is-invalid',
   },
 ];
@@ -41,14 +45,13 @@ export function getCombinations(
 ) {
   let result: Array<Object> = [];
   for (const argumentValue of argumentValues) {
-    const labelExtension: string = `[ ${argumentName} : ${argumentValue} ] `;
 
     result = [
       ...result,
       ...combinations.map(c => ({
         ...c,
         [argumentName]: argumentValue,
-        label: `${labelExtension}  ${c.label}`,
+
       })),
     ];
   }